requirements.txt
# Fuzzing and binary analysis
pwntools==4.11.0
angr==9.2.77
capstone==5.0.0.post1
keystone-engine==0.9.2
ropper==1.13.8

# AI and API integration
openai>=1.3.0,<2.0.0
# google-generativeai==0.3.0  # Commented out - replaced with local Ollama
requests==2.31.0
python-dotenv==1.0.0

# Utilities
colorama==0.4.6
tqdm==4.66.1
tabulate==0.9.0
click==8.1.7

# Binary analysis
pefile==2023.2.7
pyelftools==0.30
hexdump==3.3

# Debugging
pygdbmi==0.11.0.0
anthropic

# Local AI requirements
# Note: No additional Python packages needed for Ollama - it uses REST API